Adding final example files
The last two files added are what generates the tree view in my application. WE CAN PROBABLY USE THIS TO REFLECT THE CATEGORIES TREE FROM OSERIES.

This also shows examples of how we can call functions from the other files in the applications and pass info back and forth to the different classes in the application. Formatting and syntax from these files can be helpful to us.