DB Examples and UI Examples (#2)
* Adding additional files

Files to serve as example code we can pull from to boost this project.
• main_example.py is a renamed main.py and is used to launch the DB Selection UI which in turn launches the application handled by ui_manager.py (previously uploaded). Rename this to main.py
• delete_team_dialog.py has examples for how to spawn and use pop up dialog boxes with a drop down in it that will stay updated based on database records.
• xlsx_load_ui.py can be used in combination with files within the excel_management directory in case we need to be able to import records from excel (this portion of the application was not fully working so use at your own risk).

* Added Excel Mgmt directory

This is examples for using excel loading with python. There's an example test sheet. These files are not fully functional.

* Adding final example files

The last two files added are what generates the tree view in my application. WE CAN PROBABLY USE THIS TO REFLECT THE CATEGORIES TREE FROM OSERIES.

This also shows examples of how we can call functions from the other files in the applications and pass info back and forth to the different classes in the application. Formatting and syntax from these files can be helpful to us.